# Swagger/Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JWT",
    "version": "v1"
  },
  "paths": {
    "/api/ServiceAuftrag": {
      "get": {
        "tags": [
          "ServiceAuftrag"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceAuftragDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceAuftragDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceAuftragDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ServiceAuftrag"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceAuftragDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceAuftragDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceAuftragDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "ServiceAuftrag"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ServiceAuftrag/{id}": {
      "get": {
        "tags": [
          "ServiceAuftrag"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceAuftragDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceAuftragDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceAuftragDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ServiceAuftrag"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Token/login": {
      "post": {
        "tags": [
          "Token"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginDTO": {
        "type": "object",
        "properties": {
          "Benutzer_Name": {
            "type": "string",
            "nullable": true
          },
          "Benutzer_Passwort": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Priorität": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "ServiceAuftragDTO": {
        "type": "object",
        "properties": {
          "Auftrag_Id": {
            "type": "integer",
            "format": "int32"
          },
          "Auftrag_Dienstleistung": {
            "type": "string",
            "nullable": true
          },
          "Auftrag_Priorität": {
            "$ref": "#/components/schemas/Priorität"
          },
          "Auftrag_KundenName": {
            "type": "string",
            "nullable": true
          },
          "Auftrag_KundenEmail": {
            "type": "string",
            "nullable": true
          },
          "Auftrag_KundenTelefon": {
            "type": "integer",
            "format": "int32"
          },
          "Auftrag_status": {
            "$ref": "#/components/schemas/Status"
          },
          "Auftrag_CreateDate": {
            "type": "string",
            "format": "date-time"
          },
          "Auftrag_PickUpDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Status": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      }
    }
  }
}